edit countup component
